Update sample name
