+++ src/MAUI/Maui.Samples/Samples/Layers/AddClusteringFeatureReductionToAPointFeatureLayer/addclusteringfeaturereductiontoapointfeaturelayer.jpg
