+++ src/Samples.Shared/Resources/FeaturedSamples.xml
@@ -11,10 +11,10 @@
 			<Sample>CreateAndEditGeometries</Sample>
 		</Category>
 		<Category name="Layers">
-			<Sample>AddCustomDynamicEntityDataSource</Sample>
+      <Sample>AddClusteringFeatureReductionToAPointFeatureLayer</Sample>
+      <Sample>AddCustomDynamicEntityDataSource</Sample>
 			<Sample>AddDynamicEntityLayer</Sample>
 			<Sample>AddVectorTiledLayerFromCustomStyle</Sample>
-			<Sample>DefineClusteringFeatureReduction</Sample>
 			<Sample>DisplayPointsUsingClusteringFeatureReduction</Sample>
 			<Sample>DisplayRouteLayer</Sample>
 		</Category>
