+++ src/WPF/WPF.Viewer/Samples/Layers/AddClusteringFeatureReductionToAPointFeatureLayer/AddClusteringFeatureReductionToAPointFeatureLayer.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGIS.WPF.Samples.DefineClusteringFeatureReduction.DefineClusteringFeatureReduction"
+﻿<UserControl x:Class="ArcGIS.WPF.Samples.AddClusteringFeatureReductionToAPointFeatureLayer.AddClusteringFeatureReductionToAPointFeatureLayer"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013">
@@ -12,7 +12,7 @@
                         Padding="5"
                         Click="DrawClustersButton_Clicked"
                         Content="Draw clusters" />
-                <StackPanel x:Name="DefineClusteringFeatureReductionOptions"
+                <StackPanel x:Name="AddClusteringFeatureReductionToAPointFeatureLayerOptions"
                             Orientation="Vertical"
                             Visibility="Collapsed">
                     <TextBlock Margin="0,0,0,10"
