+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Layers/AddClusteringFeatureReductionToAPointFeatureLayer/AddClusteringFeatureReductionToAPointFeatureLayer.xaml.cs
@@ -1,10 +1,10 @@
-﻿// Copyright 2023 Esri.
+// Copyright 2023 Esri.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at: http://www.apache.org/licenses/LICENSE-2.0
 //
-// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
+// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an 
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific 
 // language governing permissions and limitations under the License.
 
 using Esri.ArcGISRuntime.ArcGISServices;
@@ -17,26 +17,26 @@ using Esri.ArcGISRuntime.Reduction;
 using Esri.ArcGISRuntime.Symbology;
 using Esri.ArcGISRuntime.Toolkit.UI.Controls;
 using Esri.ArcGISRuntime.UI.Controls;
+using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml.Controls;
 using System.Linq;
 using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
 
-namespace ArcGIS.WPF.Samples.DefineClusteringFeatureReduction
+namespace ArcGIS.WinUI.Samples.AddClusteringFeatureReductionToAPointFeatureLayer
 {
     [ArcGIS.Samples.Shared.Attributes.Sample(
-        name: "Define clustering feature reduction",
+        name: "Add clustering feature reduction to a point feature layer",
         category: "Layers",
         description: "Add client side feature reduction on a point feature layer that is not pre-configured with clustering.",
         instructions: "Tap the `Draw clusters` button to set new feature reduction object on the feature layer. Interact with the controls to customize clustering feature reduction properties. Tap on any clustered aggregate geoelement to see the cluster feature count and aggregate fields in the popup.",
         tags: new[] { "aggregate", "bin", "cluster", "group", "merge", "normalize", "popup", "reduce", "renderer", "summarize" })]
     [ArcGIS.Samples.Shared.Attributes.OfflineData("aa44e79a4836413c89908e1afdace2ea")]
-    public partial class DefineClusteringFeatureReduction
+    public partial class AddClusteringFeatureReductionToAPointFeatureLayer
     {
         private FeatureLayer _layer;
         private ClusteringFeatureReduction _clusteringFeatureReduction;
 
-        public DefineClusteringFeatureReduction()
+        public AddClusteringFeatureReductionToAPointFeatureLayer()
         {
             InitializeComponent();
             _ = Initialize();
@@ -57,6 +57,13 @@ namespace ArcGIS.WPF.Samples.DefineClusteringFeatureReduction
 
             // Set the initial viewpoint to Zurich, Switzerland.
             await MyMapView.SetViewpointAsync(new Viewpoint(47.38, 8.53, 8e4));
+
+            // Hide and nullify an opened popup when user taps screen.
+            PopupBackground.Tapped += (sender, args) =>
+            {
+                PopupBackground.Visibility = Visibility.Collapsed;
+                PopupViewer.PopupManager = null;
+            };
         }
 
         private void CreateCustomFeatureReduction()
@@ -116,7 +123,7 @@ namespace ArcGIS.WPF.Samples.DefineClusteringFeatureReduction
         }
 
         #region EventHandlers
-        private void DisplayLabelsCheckBox_Checked(object sender, RoutedEventArgs e)
+        private void DisplayLabelsCheckBox_CheckedChanged(object sender, RoutedEventArgs e)
         {
             if ((bool)(sender as CheckBox).IsChecked)
             {
@@ -135,7 +142,7 @@ namespace ArcGIS.WPF.Samples.DefineClusteringFeatureReduction
 
         }
 
-        private void UpdateClusteringProperties(object sender, RoutedPropertyChangedEventArgs<double> e)
+        private void UpdateClusteringProperties(object sender, Microsoft.UI.Xaml.Controls.Primitives.RangeBaseValueChangedEventArgs e)
         {
             if (_layer != null)
             {
@@ -154,30 +161,22 @@ namespace ArcGIS.WPF.Samples.DefineClusteringFeatureReduction
             if (result.Popups.Count == 0) return;
 
             // Set the popup and make it visible.
-            PopupViewer.Popup = result.Popups.FirstOrDefault();
+            PopupViewer.PopupManager = new PopupManager(result.Popups.FirstOrDefault());
             PopupBackground.Visibility = Visibility.Visible;
         }
-
-        private void DrawClustersButton_Clicked(object sender, RoutedEventArgs e)
+        private void DrawClustersButton_Click(object sender, RoutedEventArgs e)
         {
             // Create a new clustering feature reduction.
             CreateCustomFeatureReduction();
 
             // Show the feature reduction's clustering options.
-            DefineClusteringFeatureReductionOptions.Visibility = Visibility.Visible;
+            AddClusteringFeatureReductionToAPointFeatureLayerOptions.Visibility = Visibility.Visible;
 
             // Add an event handler for tap events on the map view.
             MyMapView.GeoViewTapped += MyMapView_GeoViewTapped;
 
             // Hide the draw clusters button.
             DrawClustersButton.Visibility = Visibility.Collapsed;
-        }
-
-        // Hide and nullify the opened popup when user left clicks.
-        private void PopupBackground_MouseLeftButtonDown(object sender, System.Windows.Input.MouseButtonEventArgs e)
-        {
-            PopupBackground.Visibility = Visibility.Collapsed;
-            PopupViewer.Popup = null;
         }
         #endregion
     }
